refactor: extract playhead mask logic, fix mousemove yield bug
